+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [3.1.0] - 2020-04-09
 ### Added
+- 500 static page.
 
 ### Changed
+- `next@10.1.3`
 
 ### Fixed