+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2020-04-09
+### Added
+- Google Analytics v4 tracking script.
+- scripts folder containing scripts used in the application.
+- 500 static page.
+
+### Changed
+- refactor: uses dynamic import for loading scripts without SSR.
+- `next@10.1.3`
+
+### Fixed
+- `/static-page` endpoint. [RW-19](https://vizzuality.atlassian.net/browse/RW-19?focusedCommentId=10534)
+
+### Removed
+
 ## [3.0.7] - 2020-04-06
 ### Added
 - 404 static page.